+++ extensions/google-chrome/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Google Chrome Changelog
 
+## [Add Exclude Search Feature] - 2025-11-03
+
+- Add support for excluding search terms using "/" prefix
+- Users can now filter out unwanted results by prefixing terms with "/"
+- Example: "raycast /firefox" finds items containing "raycast" but not "firefox"
+- Works across all search functions: tabs, history, bookmarks, and search all
+
 ## [Improvements] - 2025-10-13
 
 - Fix weird issue with a constant not working (ref: [#22053](https://github.com/raycast/extensions/issues/22053), [#22026](https://github.com/raycast/extensions/issues/22026), [#21966](https://github.com/raycast/extensions/issues/21966), [#21973](https://github.com/raycast/extensions/issues/21973), [#21928](https://github.com/raycast/extensions/issues/21928))